Move most context API calls to use a lookup based on function name rather than hard-coding the URLs
