+++ src/murfey/server/api/session_info.py
@@ -28,6 +28,7 @@ from murfey.util.db import (
     Tilt,
     TiltSeries,
 )
+from murfey.util.models import Visit
 
 logger = getLogger("murfey.server.api.session_info")
 
@@ -36,27 +37,6 @@ router = APIRouter(
     dependencies=[Depends(validate_token)],
     tags=["session info"],
 )
-
-
-class Visit(BaseModel):
-    start: datetime
-    end: datetime
-    session_id: int
-    name: str
-    beamline: str
-    proposal_title: str
-
-    def __repr__(self) -> str:
-        return (
-            "Visit("
-            f"start='{self.start:%Y-%m-%d %H:%M}', "
-            f"end='{self.end:%Y-%m-%d %H:%M}', "
-            f"session_id='{self.session_id!r}',"
-            f"name={self.name!r}, "
-            f"beamline={self.beamline!r}, "
-            f"proposal_title={self.proposal_title!r}"
-            ")"
-        )
 
 
 @router.get("/instruments/{instrument_name}/visits_raw", response_model=List[Visit])
